[dsbx] refactor: split manifest.v1 shape into dependency-free manifest_types
The manifest shape (interfaces, DB_NAME_REGEX, RESERVED_OBJECT_KEYS, error kinds) moves to
manifest_types.ts, a file with zero imports. It becomes the single authored definition of
manifest.v1: front type-imports it and asserts its zod mirror against it, so a shape change
here breaks front's typecheck instead of drifting silently. manifest.ts re-exports
everything it previously exported; no behavior change.
